backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pymongo==4.7.2
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.1.3
python-jose==3.3.0
motor==3.4.0